requirements.txt
# You can optionally pin to a version of the Runway Model SDK like: "runway-python==0.1.0"
runway-python==0.4.0
Pillow==6.0.0
numpy==1.17.3
Keras==2.3.1
